+++ core/android/src/main/kotlin/com/moengage/flutter/Constants.kt
@@ -30,5 +30,9 @@ const val METHOD_NAME_UPDATE_SDK_STATE = "updateSdkState"
 const val METHOD_NAME_ON_ORIENTATION_CHANGED = "onOrientationChanged"
 const val METHOD_NAME_UPDATE_DEVICE_IDENTIFIER_TRACKING_STATUS =
     "updateDeviceIdentifierTrackingStatus"
+const val METHOD_NAME_SETUP_NOTIFICATION_CHANNEL = "setupNotificationChannels"
+const val METHOD_NAME_NAVIGATE_TO_SETTINGS = "navigateToSettings"
+const val METHOD_NAME_REQUEST_PUSH_PERMISSION = "requestPushPermission"
+const val METHOD_NAME_PERMISSION_RESPONSE = "permissionResponse"
 
 const val KEY_TYPE = "type"