+++ core/android/src/main/kotlin/com/moengage/flutter/MoEngageFlutterPlugin.kt
@@ -97,6 +97,10 @@ class MoEngageFlutterPlugin : FlutterPlugin, MethodCallHandler {
                 METHOD_NAME_ON_ORIENTATION_CHANGED -> onOrientationChanged()
                 METHOD_NAME_UPDATE_DEVICE_IDENTIFIER_TRACKING_STATUS ->
                     updateDeviceIdentifierTrackingStatus(call)
+                METHOD_NAME_SETUP_NOTIFICATION_CHANNEL -> setupNotificationChannels()
+                METHOD_NAME_NAVIGATE_TO_SETTINGS -> navigateToSettings()
+                METHOD_NAME_REQUEST_PUSH_PERMISSION -> requestPushPermission()
+                METHOD_NAME_PERMISSION_RESPONSE -> permissionResponse(call)
                 else -> Logger.print(LogLevel.ERROR) {
                     "$tag onMethodCall() : No mapping for this" +
                             " method."
@@ -302,6 +306,42 @@ class MoEngageFlutterPlugin : FlutterPlugin, MethodCallHandler {
             pluginHelper.deviceIdentifierTrackingStatusUpdate(context, payload)
         } catch (t: Throwable) {
             Logger.print(LogLevel.ERROR, t) { "$tag updateDeviceIdentifierTrackingStatus() : " }
+        }
+    }
+
+    private fun setupNotificationChannels() {
+        try {
+            pluginHelper.setUpNotificationChannels(context)
+        } catch (t: Throwable) {
+            Logger.print(LogLevel.ERROR, t) { "$tag setupNotificationChannel() :" }
+        }
+    }
+
+    private fun navigateToSettings() {
+        try {
+            pluginHelper.navigateToSettings(context)
+        } catch (t: Throwable) {
+            Logger.print(LogLevel.ERROR, t) { "$tag navigateToSettings() :" }
+        }
+    }
+
+    private fun requestPushPermission() {
+        try {
+            pluginHelper.requestPushPermission(context)
+        } catch (t: Throwable) {
+            Logger.print(LogLevel.ERROR, t) { "$tag requestPushPermission() :" }
+        }
+    }
+
+    private fun permissionResponse(methodCall: MethodCall) {
+        try {
+            Logger.print { "$tag permissionResponse() : Arguments: ${methodCall.arguments}" }
+            if (methodCall.arguments == null) return
+            val payload = methodCall.arguments.toString()
+            Logger.print { "$tag selfHandledCallback() : Payload: $payload" }
+            pluginHelper.permissionResponse(context, payload)
+        } catch (t: Throwable) {
+            Logger.print(LogLevel.ERROR, t) { "$tag permissionResponse() :" }
         }
     }
 }