+++ core/CHANGELOG.md
@@ -1,6 +1,11 @@
 # MoEngage Flutter Plugin
 
-### Next Release *()*
+=======
+## Next Release
+- Android
+  - Build Configuration Updates
+    - Compile SDK Version - 31
+    - Target SDK version - 31
 - iOS
   - Deprecated API
 
@@ -8,8 +13,12 @@
 |:------------------------------------------------------------------------------------------------------------------------------------:|:---------------------------------------------------------------------------------------------------------------------------------------------------:|
 | initializeDefaultInstance(_ config: MOSDKConfig, sdkState: Bool = true, launchOptions: [UIApplication.LaunchOptionsKey: Any]? = nil) | initializeDefaultInstance(config: MOSDKConfig, sdkState: MoEngageSDKState = .enabled, launchOptions: [UIApplication.LaunchOptionsKey: Any]? = nil)  |
 
+### 5.1.0
+- Android
+  - Support for Android 13 notification permission.
+  - Android Gradle Plugin version updated to `7.3.1`
+  - Gradle version updated to `7.4`
 
-  
 ## 27.09.2022
 
 ### 5.0.0