+++ inbox/CHANGELOG.md
@@ -1,5 +1,14 @@
 # MoEngage Flutter Inbox Plugin
 
+### Next Release
+
+### 4.1.0
+- Android
+    - Android Gradle Plugin version updated to `7.3.1`
+    - Gradle version updated to `7.4`
+    - Compile SDK Version - 31
+    - Target SDK version - 31
+
 ## 27.09.2022
 
 ### 4.0.0